requirements.txt
requests~=2.32.3
dotenv~=0.9.9
python-dotenv~=1.0.1
psycopg2-binary==2.9.10
fastapi~=0.115.11
PyYAML~=6.0.2
httpx~=0.28.1
SQLAlchemy~=2.0.38
pydantic~=2.10.6
psycopg2~=2.9.10